Delete README.md
This project queries a relational database on USDA to review the nutrition value of food items in the United State and are required to provide answers to the following questions and to come up with insights that would help the department to understand nutrients available to the American people rentals to answer specific questions. Files included are the queries, report detailing the visualized insights in a PDF format, the entity relationship diagram (ERD) for the database and this README file.

Do find it interesting.